Add missing carousel JS files
Owl carousel had 3 includes on the template, but only one of the files was in the repo. Sourced the missing files for the version matching the one file there was.

The targets of owl carousel in custom.js dont appear to exist in this HTML template, so these includes and those parts of custom.js could be extracted and removed from this page. Then included only when utilised.