Task 6 - Change new arrivals to 3 columns
Used css grid to make 3 columns, going down to 2 for mobile. (Grid could be used to make it more responsive without need for media queries using the auto-fill/fit methods).

Amended layout of stars and prices to align better and be more responsive.

Potentially the design of these items could be made to match the design of the items in the new collection items with reguards to title/image/price/star layout, or visa-versa.